pyproject.toml
[tool.poetry]
name = "emojibot"
version = "0.1.0"
description = "a bot that helps you quickly import emojis from a message or an image."
authors = ["j1nxie <rylieeeeexd@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.1"
python-dotenv = "^1.0.0"
jishaku = "^2.5.1"
asyncio = "^3.4.3"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
jishaku = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
